requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
httpx==0.28.0
openai==1.57.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
python-multipart==0.0.18
python-dotenv==1.0.1
aiofiles==24.1.0
